#21 feat(be): MinutePriceResponse 구현
- 응답 받은 당일일분봉 데이터를 가공해 클라이언트에게 전달하는 Dto
